backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
mediapipe>=0.10.30
tensorflow>=2.15.0
opencv-python>=4.10.0
numpy>=1.26.0
pillow>=10.0.0
scikit-learn>=1.4.0
